The fourth part of the InternetShop

+++ src/main/java/mate/academy/internetshop/Main.java
@@ -1,6 +1,5 @@
 package mate.academy.internetshop;
 
-import java.math.BigDecimal;
 import mate.academy.internetshop.dao.Storage;
 import mate.academy.internetshop.lib.Injector;
 import mate.academy.internetshop.model.Product;
@@ -15,7 +14,7 @@ public class Main {
         Storage.products.forEach(System.out::println);
         System.out.println();
         productService.delete(2L);
-        Product product4 = new Product("Lens X30.3", BigDecimal.valueOf(30.3));
+        Product product4 = new Product("Lens X30.3", 30.3);
         product4.setProductId(3L);
         productService.update(product4);
         System.out.println();
@@ -23,9 +22,9 @@ public class Main {
     }
 
     private static void initializeDb(ProductService productService) {
-        Product product1 = new Product("Lens X10", BigDecimal.valueOf(10));
-        Product product2 = new Product("Lens X20", BigDecimal.valueOf(20));
-        Product product3 = new Product("Lens X30", BigDecimal.valueOf(30));
+        Product product1 = new Product("Lens X10", 10);
+        Product product2 = new Product("Lens X20", 20);
+        Product product3 = new Product("Lens X30", 30);
         productService.create(product1);
         productService.create(product2);
         productService.create(product3);
