+++ src/main/java/mate/academy/internetshop/model/Product.java
@@ -7,9 +7,9 @@ public class Product {
     private String name;
     private BigDecimal price;
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, double price) {
         this.name = name;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 
     @Override
